api/requirements.txt
django-extensions
django-log-request-id
django-micro
djangorestframework
drf-spectacular
drf-spectacular-sidecar
elastic-apm
faker
freezegun
ipdb
ipython
python-json-logger
